pyproject.toml
[tool.poetry]
name = "langchain-001"
version = "0.1.0"
description = ""
authors = ["Jun Terayama <pam130901@gmail.com>"]

[tool.poetry.dependencies]
python = "3.11.3"
python-dotenv = "^1.0.1"
openai = "^1.12.0"
langchain = "0.1.6"
langchain-openai = "^0.0.6"
unstructured = "^0.12.4"
slack-bolt = "^1.18.1"
gradio = "3.48.0"
chromadb = "0.3.29"
langchain-experimental = "0.0.39"
duckduckgo-search = "^4.5.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
